conditionally include memrchr()
