+++ cmake/Modules/EttercapLibCheck.cmake
@@ -173,6 +173,7 @@ endif(NOT HAVE_STRLCAT_FUNCTION OR NOT HAVE_STRLCPY_FUNCTION)
 CHECK_FUNCTION_EXISTS(strsep HAVE_STRSEP)
 CHECK_FUNCTION_EXISTS(strcasestr HAVE_STRCASESTR)
 CHECK_FUNCTION_EXISTS(memmem HAVE_MEMMEM)
+CHECK_FUNCTION_EXISTS(memrchr HAVE_MEMRCHR)
 CHECK_FUNCTION_EXISTS(basename HAVE_BASENAME)
 CHECK_FUNCTION_EXISTS(strndup HAVE_STRNDUP)
 CHECK_FUNCTION_EXISTS(inet_aton HAVE_INET_ATON)
