+++ src/CMakeLists.txt
@@ -163,6 +163,9 @@ endif(NOT HAVE_STRCASESTR)
 if(NOT HAVE_MEMMEM)
     set(EC_SRC ${EC_SRC} missing/memmem.c)
 endif(NOT HAVE_MEMMEM)
+if(NOT HAVE_MEMRCHR)
+    set(EC_SRC ${EC_SRC} missing/memrchr.c)
+endif(NOT HAVE_MEMRCHR)
 if(NOT HAVE_BASENAME)
     set(EC_SRC ${EC_SRC} missing/basename.c)
 endif(NOT HAVE_BASENAME)
