+++ src/missing/memrchr.c
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) 2001 Alberto Ornaghi <alor@users.sourceforge.net>
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. The name of the author may not be used to endorse or promote products
+ *    derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES,
+ * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY
+ * AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL
+ * THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include <ec.h>
+
+void *memrchr(const void *s, u_char c, size_t n)
+{ 
+
+/* 
+ *  DESCRIPTION
+ *        The  memrchr() function is like the memchr() function, except
+ *        that it searches backward for the first instance of c from the
+ *        end of the n bytes pointed to by s instead of forward from the
+ *        beginning.
+ *  
+ *  RETURN VALUE
+ *       The memrchr() functions return a pointer to the matching byte or
+ *       NULL if the  character does not occur in the given memory area.
+ */
+
+   const u_char *cp;
+
+   if (n != 0) {
+      cp = (u_char*)s + n;
+      do {
+         if (*(--cp) == (u_char)c)
+            return (void*)cp;
+      } while (--n != 0);
+   }
+
+   return NULL;
+}
+
